Integrate module gen/programs into driver/tool
